added (float) to divide functions

+++ src/color/color_math_basics.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   color_math.c                                       :+:    :+:            */
+/*   color_math_basics.c                                :+:    :+:            */
 /*                                                     +:+                    */
 /*   By: jaberkro <jaberkro@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/26 14:49:47 by jaberkro      #+#    #+#                 */
-/*   Updated: 2022/10/26 14:52:24 by jaberkro      ########   odam.nl         */
+/*   Updated: 2022/11/04 16:11:20 by jaberkro      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,9 +46,9 @@ t_color	multiply_color_float(t_color color, float t)
 {
 	t_color	result;
 
-	result.r = color.r * t;
-	result.g = color.g * t;
-	result.b = color.b * t;
+	result.r = (float)color.r * t;
+	result.g = (float)color.g * t;
+	result.b = (float)color.b * t;
 	return (result);
 }
 
@@ -56,8 +56,8 @@ t_color	divide_color_with_float(t_color color, float t)
 {
 	t_color	result;
 
-	result.r = color.r / t;
-	result.g = color.g / t;
-	result.b = color.b / t;
+	result.r = (float)color.r / t;
+	result.g = (float)color.g / t;
+	result.b = (float)color.b / t;
 	return (result);
 }
